requirements.txt
flask
waitress
pytelegrambotapi
telebot
pyngrok
requests
python-dotenv
colorama